+++ bindings/pydeck/pydeck/frontend_semver.py
@@ -1,0 +1,1 @@
+DECKGL_SEMVER = "^8.0.0"